+++ examples/blast-with-force-and-skipping/0-start-log.sh
@@ -3,4 +3,4 @@
 echo "Running (fake) BLAST pipeline at $(date)" > pipeline.log
 echo "Input FASTA is $@" >> pipeline.log
 echo "SP_FORCE is $SP_FORCE" >> pipeline.log
-echo "SP_SIMULATE is $SP_SIMULATE" >> pipeline.log
+echo "SP_SKIP is $SP_SKIP" >> pipeline.log