+++ examples/blast-with-force-and-skipping/blast.sh
@@ -3,9 +3,9 @@
 echo >> pipeline.log
 echo "In $(basename $0) $@" >> pipeline.log
 echo "SP_FORCE is $SP_FORCE" >> pipeline.log
-echo "SP_SIMULATE is $SP_SIMULATE" >> pipeline.log
+echo "SP_SKIP is $SP_SKIP" >> pipeline.log
 
-# $1 = "--query" (to simulate running BLAST), which we just ignore.
+# $1 = "--query" (to simulate calling BLAST), which we just ignore.
 # $3 = "--outfmt" (ditto).
 
 # $2 is given to us by run-blast.sh (it's one of the x?? FASTA files). Pull
@@ -15,9 +15,9 @@ queryId=$(head -n 1 $task | cut -c2-)
 
 out=$task.blast-out
 
-if [ $SP_SIMULATE = "0" ]
+if [ $SP_SKIP = "0" ]
 then
-    echo "This is not a simulation." >> pipeline.log
+    echo "Not skipping." >> pipeline.log
 
     if [ -f $out ]
     then
@@ -35,7 +35,7 @@ then
         echo "$RANDOM subject-$RANDOM $queryId" > $out
     fi
 else
-    echo "This is a simulation. Will re-emit task name ($task)." >> pipeline.log
+    echo "Skipping. Will re-emit task name ($task)." >> pipeline.log
 fi
 
 echo "TASK: $task"